6-69~70. Assignment - JUnit Dependency Injection
