+++ src/test/java/com/udemy/junit_lecture_1/testJavaWithJunit/model/PersonTest.java
@@ -2,11 +2,7 @@ package com.udemy.junit_lecture_1.testJavaWithJunit.model;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.RepeatedTest;
-import org.junit.jupiter.api.RepetitionInfo;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestInfo;
 
 import com.udemy.junit_lecture_1.testJavaWithJunit.ModelTests;
 
@@ -34,17 +30,5 @@ class PersonTest implements ModelTests {
             () -> assertEquals("Joe", person.getFirstName(), "First Name Failed"),
             () -> assertEquals("Buck", person.getLastName(), "Last Name Failed")
         );
-    }
-
-    // {displayName}이 name의 가장 앞에 오는 경우 run view 에서는 생략되어 보인다.
-    @RepeatedTest(value = 10, name = "{displayName} : {currentRepetition} - {totalRepetitions}")
-    @DisplayName("My Repeated Test")
-    void myRepeatedTest() {
-        // TODO - impl
-    }
-
-    @RepeatedTest(5)
-    void myRepeatedTestWithDI(TestInfo testInfo, RepetitionInfo repetitionInfo) {
-        System.out.println(testInfo.getDisplayName() + ": " + repetitionInfo.getCurrentRepetition());
     }
 }
